third_party/blink/web_tests/external/wpt/performance-timeline/not-restored-reasons/abort-block-bfcache.window-expected.txt

This is a testharness.js-based test.
[FAIL] aborting a parser should block bfcache.
  assert_equals: expected "parser-aborted" but got "loading"
Harness: the test ran to completion.
